fix some warnings

+++ src/Output.h
@@ -13,7 +13,7 @@ void render_map     (sf::Uint8* pixels, Map_t* map, PlayerSet_t* PlayerSet);
 void make_screenshot(sf::RenderWindow* window, const char* output_file);
 
 void print_state_info(sf::RenderWindow* window, sf::Text* POS_Text, sf::Text* FPS_Text,
-                      char* pos_string, int len_pos_string, char* fps_string, int len_fps_string,
+                      char* pos_string, size_t len_pos_string, char* fps_string, size_t len_fps_string,
                       std::chrono::_V2::steady_clock::time_point clock_begin,
                       std::chrono::_V2::steady_clock::time_point clock_end,
                       PlayerSet_t* PlayerSet, double* old_fps);
